requirements.txt
argcomplete==1.12.0
astroid==2.4.2
attr==0.3.1
attrs==20.2.0
backoff==1.8.0
ciso8601==2.1.3
click==7.1.2
colorama==0.4.3
isort==5.5.2
jsonschema==2.6.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.4
pipx==0.15.5.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2018.4
simplejson==3.11.1
singer-python==5.9.0
six==1.15.0
toml==0.10.1
userpath==1.4.1
wrapt==1.12.1